+++ test/exactly_lib_test/test_case_utils/file_matcher/path/z_package_suite.py
@@ -1,0 +1,14 @@
+import unittest
+
+from exactly_lib_test.test_case_utils.file_matcher.path import glob_pattern, reg_ex
+
+
+def suite() -> unittest.TestSuite:
+    return unittest.TestSuite([
+        glob_pattern.suite(),
+        reg_ex.suite(),
+    ])
+
+
+if __name__ == '__main__':
+    unittest.TextTestRunner().run(suite())